create BlocksUtil file - refactor
